+++ src/apollo/queries.js
@@ -1,3 +1,4 @@
+import { apiTimeframeOptions } from 'constants/apiTimeframeOptions'
 import gql from 'graphql-tag'
 
 const TokenFields = `
@@ -15,6 +16,33 @@ const TokenFields = `
     feesUSD
   }
 `
+const PoolFields = `
+  fragment PoolFields on Pool {
+    poolAddress
+    token0 {
+      tokenAddress
+      symbol
+      name
+      #totalValueLocked
+    }
+    token1 {
+      tokenAddress
+      symbol
+      name
+      #totalValueLocked
+    }
+    volumeToken0
+    volumeToken1
+    volumeUSD
+    totalValueLockedUSD
+    totalValueLockedETH
+    totalValueLockedToken0
+    totalValueLockedToken1
+    token0Price
+    token1Price
+    fee
+  }
+`
 
 export const TOKENS_DATA = ({ tokenIds = [] }) => {
   const tokenString = `[${tokenIds.map((token) => `"${token}"`).join(',')}]`
@@ -28,5 +56,39 @@ export const TOKENS_DATA = ({ tokenIds = [] }) => {
       }
     }
   `
+  return gql(queryString)
+}
+
+export const HISTORICAL_POOLS_DATA = ({ tokenIds = [], periods = [] }) => {
+  const tokensString = `[${tokenIds.map((token) => `"${token}",`)}]`
+  const periodString = `[${periods.map((period) => `"${period}"`).join(',')}]`
+
+  let queryString = `
+    ${PoolFields}
+    query poolsData {
+      poolsData(
+        first: 500, 
+        where: {
+          periodIn: ${periodString},
+          bothTokenAddressIn: ${tokensString},
+        }
+      ) {
+        pool {
+          ...PoolFields
+        }
+        period
+      }
+    }
+  `
+  return gql(queryString)
+}
+
+export const HISTORICAL_GLOBAL_DATA = () => {
+  const queryString = ` query jediswapFactories {
+      factoriesData {
+        ${apiTimeframeOptions.oneDay}
+        ${apiTimeframeOptions.twoDays}
+      }
+    }`
   return gql(queryString)
 }
