+++ src/pages/Pool/index.tsx
@@ -23,12 +23,12 @@ import { useUserHideClosedPositions } from 'state/user/hooks'
 import { ThemedText } from 'theme/components'
 import { LoadingRows } from './styled'
 import fetchTokenIds from 'api/fetchTokenId'
-import { getAllPools } from 'graphql/data/PoolsData'
+import { getAllPools } from 'api/PoolsData'
 import Pools from 'components/Pools'
 import { ToggleElement, ToggleWrapper } from 'components/Toggle/MultiToggle'
 import { formattedNum, formattedPercent, get2DayPercentChange, getPercentChange } from 'utils/dashboard'
 import { REWARDS_SELECTOR, STARKNET_REWARDS_API_URL } from 'constants/strkRewards'
-import { HISTORICAL_GLOBAL_DATA } from 'graphql/data/queries'
+import { HISTORICAL_GLOBAL_DATA } from 'apollo/queries'
 import { apiTimeframeOptions } from 'constants/apiTimeframeOptions'
 import { DEFAULT_CHAIN_ID, ETH_ADDRESS, NONFUNGIBLE_POOL_MANAGER_ADDRESS } from 'constants/tokens'
 import { providerInstance } from 'utils/getLibrary'
